Add MetalHostValueSource.
