+++ plugins/src/main/java/br/dev/pedrolamarao/gradle/metal/MetalService.java
@@ -7,24 +7,17 @@ import org.gradle.api.provider.Provider;
 import org.gradle.api.provider.ProviderFactory;
 import org.gradle.api.services.BuildService;
 import org.gradle.api.services.BuildServiceParameters;
-import org.gradle.process.ExecOperations;
 
 import javax.inject.Inject;
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.IOException;
-import java.io.StringReader;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * Metal service.
  */
 public abstract class MetalService implements BuildService<BuildServiceParameters.None>
 {
-    private final AtomicReference<String> host = new AtomicReference<>();
-
     private final String path;
 
     private final String target;
@@ -45,14 +38,6 @@ public abstract class MetalService implements BuildService<BuildServiceParameter
     }
 
     /**
-     * Exec operations service.
-     *
-     * @return service
-     */
-    @Inject
-    protected abstract ExecOperations getExec ();
-
-    /**
      * Provider factory service.
      *
      * @return service
@@ -67,32 +52,7 @@ public abstract class MetalService implements BuildService<BuildServiceParameter
      */
     public Provider<String> getHost ()
     {
-        return getProviders().provider(host::get).orElse(
-            getProviders().exec(exec ->
-            {
-                exec.executable( Metal.toExecutableFile(path,"clang") );
-                exec.args("-v");
-            })
-            .getStandardError().getAsText().map(buffer ->
-            {
-                String value = "unknown";
-                try (var reader = new BufferedReader( new StringReader( buffer.toString() ) ) )
-                {
-                    String line;
-                    while ((line = reader.readLine()) != null) {
-                        if (line.startsWith("Target")) {
-                            final var split = line.split(":");
-                            if (split.length != 2) continue;
-                            value = split[1].trim();
-                            break;
-                        }
-                    }
-                }
-                catch (IOException e) { throw new GradleException("failed parsing clang -v output", e); }
-                host.set(value);
-                return value;
-            })
-        );
+        return getProviders().of(MetalHostValueSource.class,spec -> spec.parameters(it -> it.getPath().set(getPath())));
     }
 
     /**
@@ -130,21 +90,21 @@ public abstract class MetalService implements BuildService<BuildServiceParameter
      * @param name    core name
      * @return        file name
      */
-    public String archiveFileName (String target, String name)
+    public static String archiveFileName (String target, String name)
     {
         final var prefix = archiveFilePrefix(target);
         final var suffix = archiveFileSuffix(target);
         return "%s%s%s".formatted(prefix,name,suffix);
     }
 
-    String archiveFilePrefix (String target)
+    static String archiveFilePrefix (String target)
     {
         final var parts = target.split("-");
         if (contains(parts,"msvc")) return "";
         else return "lib";
     }
 
-    String archiveFileSuffix (String target)
+    static String archiveFileSuffix (String target)
     {
         final var parts = target.split("-");
         if (contains(parts,"msvc")) return ".lib";
@@ -169,19 +129,19 @@ public abstract class MetalService implements BuildService<BuildServiceParameter
      * @param name    core name
      * @return        file name
      */
-    public String executableFileName (String target, String name)
+    public static String executableFileName (String target, String name)
     {
         final var prefix = executableFilePrefix(target);
         final var suffix = executableFileSuffix(target);
         return "%s%s%s".formatted(prefix,name,suffix);
     }
 
-    String executableFilePrefix (String target)
+    static String executableFilePrefix (String target)
     {
         return "";
     }
 
-    String executableFileSuffix (String target)
+    static String executableFileSuffix (String target)
     {
         final var parts = target.split("-");
         if (contains(parts,"windows")) return ".exe";
